feat: Add pool index in CLPoolManager
If you want to query v4 pool information, you need to know the poolId, which is very long and not user-friendly.
So i suggest we can record the pool index which start from 1,  we just need to remember the pool index, then we can use index to query the pool information easily.