feat: Remove unnecessary poolManagerMatch modifier
We only need to check the poolManager when initialize.
Then the pool id is confirmed after initialized.
when user call other functions , like modifyLiquidity , we will call _checkPoolInitialized, so it is unnecessary to check the poolManager address again , because if they pass an invalid poolManager in PoolKey, the pool id will not be exist, will get PoolNotInitialized error.